+++ Realm/Tests/ObjectTests.m
@@ -141,6 +141,21 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
 }
 @end
 
+@interface StringSubclassObject : StringObject
+@property NSString *stringCol2;
+@end
+
+@implementation StringSubclassObject
+@end
+
+@interface StringLinkObject : RLMObject
+@property StringObject *stringObjectCol;
+@property RLMArray<StringObject> *stringObjectArrayCol;
+@end
+
+@implementation StringLinkObject
+@end
+
 @interface ReadOnlyPropertyObject ()
 @property (readwrite) int readOnlyPropertyMadeReadWriteInClassExtension;
 @end
@@ -422,6 +437,32 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
     XCTAssertEqualObjects(row2.mixedCol, @2,            @"row2.mixedCol");
 }
 
+- (void)testObjectSubclass {
+    RLMRealm *realm = [RLMRealm defaultRealm];
+    [realm beginWriteTransaction];
+    [StringObject createInDefaultRealmWithObject:@[@"string"]];
+    StringSubclassObject *obj = [StringSubclassObject createInDefaultRealmWithObject:@[@"string", @"string2"]];
+
+    // ensure property ordering
+    XCTAssertEqualObjects([obj.objectSchema.properties[0] name], @"stringCol");
+    XCTAssertEqualObjects([obj.objectSchema.properties[1] name], @"stringCol2");
+
+    [realm commitWriteTransaction];
+
+    // ensure creation in proper table
+    RLMArray *results = StringSubclassObject.allObjects;
+    XCTAssertEqual(1U, results.count);
+    XCTAssertEqual(1U, StringObject.allObjects.count);
+
+    // ensure exceptions on when using polymorphism
+    [realm beginWriteTransaction];
+    StringLinkObject *linkObject = [StringLinkObject createInDefaultRealmWithObject:@[NSNull.null, @[]]];
+    XCTAssertThrows(linkObject.stringObjectCol = obj);
+    XCTAssertThrows([linkObject.stringObjectArrayCol addObject:obj]);
+    [realm commitWriteTransaction];
+}
+
+
 #pragma mark - Default Property Values
 
 - (void)testNoDefaultPropertyValues
@@ -507,7 +548,6 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
     for (NSString *key in keys) {
         NSMutableDictionary *dict = [inputValues mutableCopy];
         dict[key] = NSNull.null;
-        [dict removeObjectForKey:key];
         [DefaultObject createInRealm:realm withObject:dict];
     }
 
@@ -777,6 +817,18 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
         [RLMRealm.defaultRealm addObject:obj];
         [obj.objects addObject:obj];
     }];
+    XCTAssertNoThrow(obj.description);
+}
+
+- (void)testDeletedObjectDescription
+{
+    RLMRealm *realm = [RLMRealm defaultRealm];
+
+    [realm beginWriteTransaction];
+    EmployeeObject *obj = [EmployeeObject createInRealm:realm withObject:@[@"Peter", @30, @YES]];
+    [realm deleteObject:obj];
+    [realm commitWriteTransaction];
+
     XCTAssertNoThrow(obj.description);
 }