+++ CHANGELOG.md
@@ -7,11 +7,16 @@
 
 ### Enhancements
 
+* Support subclassing RLMObject models. Although you can now persist subclasses
+  polymorphic behavior is not supported (i.e. setting a property to an 
+  instance of its subclass).
 * Add support for sorting RLMArray properties.
 * Speed up inserting objects with `addObject:` by ~20%.
 * `readonly` properties are automatically ignored rather than having to be
   added to `ignoredProperties`.
 * Updating to core library version 0.83.1.
+* Return "[deleted object]" rather than throwing an exception when
+  `-description` is called on a deleted RLMObject.
 
 ### Bugfixes
 
@@ -19,6 +24,8 @@
   type when building for devices with Xcode 6.
 * Fix spurious notifications of new versions of Realm.
 * Fix for updating nested objects where some types do not have primary keys.
+* Fix for inserting objects from JSON with NSNull values when default values
+  should be used.
 
 0.85.0 Release notes (2014-09-15)
 =============================================================